Rename S3Client -> S3CrtClient
This gives us some room to experiment with other clients in the same
tree. This change is a pure automated refactor, no other code changes.
