+++ s3-file-connector/examples/prefetch_benchmark.rs
@@ -5,7 +5,7 @@ use std::time::Instant;
 use aws_crt_s3::common::rust_log_adapter::RustLogAdapter;
 use clap::{Arg, Command};
 use futures::executor::ThreadPool;
-use s3_client::{S3Client, S3ClientConfig};
+use s3_client::{S3ClientConfig, S3CrtClient};
 use s3_file_connector::prefetch::Prefetcher;
 use tracing_subscriber::fmt::Subscriber;
 use tracing_subscriber::util::SubscriberInitExt;
@@ -74,7 +74,7 @@ fn main() {
         throughput_target_gbps,
         part_size,
     };
-    let client = Arc::new(S3Client::new(region, config).expect("couldn't create client"));
+    let client = Arc::new(S3CrtClient::new(region, config).expect("couldn't create client"));
 
     for i in 0..iterations.unwrap_or(1) {
         let runtime = ThreadPool::builder().pool_size(1).create().unwrap();
